Excercise 4 - Perspective Decoration
Added a component for the decoration (though it could just be a div in
the CharacterEditor, I figured it's nice to be reuseable on other pages
where we might want the same effect).

It's simply a fixed positioned box. The maxwidthWrapper needs to become
position: relative otherwise the perspectivedecoration would render
above everything else (because of the fact that all static elements are
rendered first and separately from fixedones)